requirements.txt
numpy
pandas
torch
transformers
lancedb
boto3
Pillow
tqdm
scikit-learn
colpali-engine
